Deliver the ansi_up via our starter
As it turns out, src/main/static/* from within starters is recognized as static resources as well, so we can deliver the nice little script with the starter. Being at it, we add an "enableAnsiBanners" element to our dialect that creates a script node containing the resource as well as the inline script.